Add GitHub Actions workflow (#24)
* Add GitHub Actions workflow

Adds CI workflow to run `yarn lint`, `yarn test`, and `yarn build`. `yarn flow` excluded for now as we still have many typecheck failures.

Based on:
- Default Node.js workflow
- https://github.com/actions/cache/blob/master/examples.md#node---yarn

* Prettify node.js.yml

* Configure test script to pass with no tests

Tests are currently failing as there are no tests in the codebase. This
commit adds `--passWithNoTests` to make Jest not error.

This commit should be reverted once there are tests in our codebase.

* Revert "Configure test script to pass with no tests"

This reverts commit bd57b9ea82328fb3765b91b7847eb99f51b7aae4.

* Fix lint in new files on master