Fix preprocessDataV2 to prevent enormous React timestamps
`preprocessDataV2` previously took the `ts` value of the first timeline
event, following `preprocessData`. However, this is problematic with
Chrome performance data, as Chrome sticks a bunch of metadata events
(i.e. `ph === 'M'`) with `ts === 0`. This caused the computed React
events and measures to have enormous `timestamp` values, since they were
now offset from 0 instead of being offset from the first event.

This commit fixes this issue by taking the `ts` value of the first event
with a non-zero and non-undefined `ts`.

This fix was implemented with reference to the trace event doc:
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview

1. V2 measures now appear correctly in WIP (not pushed) canvas with V2
   data.
1. CI. Tests still passing.